Move _normalize_hosts into utils
